cassita/server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
torch
diffusers
pillow
accelerate
transformers